test(osr): zoom-resize soak probe + present size diag — proves the size-gate under soak
Soak-proven on the flutter_cef side (real CEF, single-view) that the size-gated promotion
works and does NOT degrade: cycling renderScale (dpr) hammers the resize path; the
[cefdiag-resize] log shows, per present while a resize is pending, the actual composited
src dims vs the expected new-surface dims (round(logical*dpr)) and whether they match.

Verbatim soak evidence (dpr sweep): each resize logs `src=OLD exp=NEW match=false` (the
renderer's lagging frame — correctly NOT promoted) then `src=NEW exp=NEW match=true` (the
re-rastered frame — promoted). 38 match=true / 17 match=false over the run, still matching
17s in → no stick/degradation. Confirms src (view_src dims) is the true frame size (never
pool-sized), so the size-gate is sound.

- example/lib/zoom_soak_probe.dart: the soak harness (auto-cycles renderScale, ticking
  clock + fixed-proportion box make freeze/wrong-scale obvious).
- CefWebSession: [cefdiag-resize] present-size diagnostic (behind FLUTTER_CEF_DEBUG).
- CefProfileHost: explicit UInt64(0) in the F-6 ternary (the example's fresh pod surfaced
  a type-inference error the Campus build had masked).

NOTE: the prior Campus "scale-fix" builds were broken by a build-infra bug, not the fix —
`make cef-host`'s up-to-date gate keys on the flutter_cef path stamp, not main.mm's mtime,
so local native edits were silently not recompiled (old 4-byte present vs new 12-byte
parser → promotion never fired). Force a clean cef_host rebuild (rm the .flutter_cef_ref
stamp) after any native edit.

+++ packages/flutter_cef_macos/macos/Classes/CefProfileHost.swift
@@ -803,7 +803,7 @@ final class CefProfileHost {
         let nudged = c.nudgedAt != 0
         let action = LivenessProbePolicy.evaluate(
           sinceLastPresentNs: c.sinceLast, stalenessThresholdNs: livenessStalenessNs,
-          nudged: nudged, sinceNudgeNs: nudged ? (now &- c.nudgedAt) : 0,
+          nudged: nudged, sinceNudgeNs: nudged ? (now &- c.nudgedAt) : UInt64(0),
           nudgeGraceNs: livenessGraceNs)
         switch action {
         case .healthy:
